requirements.txt
alpaca-trade-api
pytz
requests
python-dateutil